Add Lyric Input Menu
Add a popup window to input lyrics into.  Doesn't allow for scrolling, since Input Fields are not playing nice with the layout system. Pressing enter also causes the input field to lose focus, which is not ideal, but I do still want focus to be lost when the user clicks elsewhere. Will look into later.

Also added buttons to the Lyric Editor to open the lyric input menu and to place a lyric; no functionality yet, however.